dns: default to always loading nm-dns-systemd-resolve
If the primary plugin is different, create an instance of the
systemd-resolve plugin alongside it to keep systemd-resolve up to date.

If for whatever reasons don't want NetworkManager to push the DNS data
it discovers to systemd-resolved, the functionality can be disabled
with:

  [main]
  dns-=systemd-resolve

Invalid configuration is ignored, not sure if there's anything better to
do instead.