device: apply static addresses immediately for DHCPv4 method
When the IPv4 method is 'auto' and there are static addresses
configured in the connection, start a DAD probe for the static
addresses and apply them immediately on success, without waiting for
DHCP to complete.

Note that if the static address is in the same subnet of the DHCP one,
when we add the DHCP address we want it to be primary and so we will
remove the static address temporarily to have the right order of
addresses.

https://bugzilla.redhat.com/show_bug.cgi?id=1369905