fixup! platform: allow setting drivers-autoprobe on SR-IOV PFs
Virtual functions in general have only one caller. Since there
is only one caller, we should put more trust into the caller
passing valid arguments.

In fact, the caller already contains the same assertion. No need
to repeat it.

(the same is often true for static function: if there are few callers,
or if it's easy to verify that all callers don't pass invalid arguments,
there is less reason for a runtime assertion).