libnm: add generic-data interface to NMSetting
libnm's NMConnection is basically a dictionary of settings ("a{sa{sv}}"
on D-Bus), and NMSetting is basically a dictionary of variants ("a{sv}"
on D-Bus).

As part of libnm they both provide useful functionality for interpreting
profiles as received from D-Bus, for example

  - nm_connection_verify() can check for consistancy
  - nm_connection_compare() can check for semantical equality
    (unlike g_variant_equal() which just compare whether two
    variants are literally equal).
  - they provide introspection data so that it is easier to
    understand which properties are available. Mostly they
    do so via GObject properties.

However, they have downsides:

- they abstract too much. While a profile can be represented just as a
GVariant, once it's converted to a NMConnection/NMSetting, the variant
is no longer accessible (why?).

- their native meta data is not sophisticated enough. Sure, there are
GObject properties. But if this value is for example of type "string",
which values are valid? How to generically convert a property to string
and back? Which is something what nmcli, keyfile and ifcfg-rh do extensively.

 - For nmcli we need all of "clients/common/nm-meta-setting-desc.c" to handle this.

 - "libnm-core/nm-keyfile.c" is less generic that it should be, and needs
   ParseInfoProperty extensively.

 - ifcfg-rh reader/writer is not generic at all, even for places where it
   could be and doesn't have to follow initscripts API.

 - VPN handling of profile basically reduces everything to a dump string
   dictionary. The user experience with nmcli is very bad for VPN
   profiles.

- whenever we add a new property, or even a new NMSetting instance, we
need an incredible amount of code, just to implement this. Something
that should be as simple as having one general NMGenericSetting, which
has internally a vtable of calls for verify(), normalize(), compare(),
etc.

This patch doesn't solve these problems. It's just a small step of
making NMSetting more generic. By exposing the GVariant values directly.

As it is currently intended, only subclasses are supposed to set the
gendata values. That means, for existing NMSetting types, nothing
changes, including their GVariant values will not be accessible via
this API. But new properties can make use of this API, to implement
propeties that are easier to extend.

This is done, because I am about to add ethtool options. Ethtool has a
large number of possible options, and preferably we will successively
add more and more. Adding and supporting a new ethtool options must
be low effort. That does not scale with the previous model.

This does not mean, that derived classes cannot (or shall not) add
GObject properties or C functions for direct access. On the contrary,
these might be very important for a more convenient use. However,
implementing them should be possible more generically.