libnm: handle setting registration also via nm-meta-setting.h
Previously, each (non abstract) NMSetting class had to register
its name and priority via _nm_register_setting().

That metadata was partly redundant to nm-meta-setting.h and
NMMetaSettingInfo. Merge them.

Note, that previously, _nm_register_setting() was private API. That
means, no user could subclass a working NMSetting instance. The same
is still true: NMMetaSettingInfo is internal API and users cannot access
it to create their own NMSetting subclasses. But that is almost desired.
libnm is not designed, to be extensible via subclassing, nor is it
clear why that would be a useful thing to do.

This change has the following benefits:

- the main benefit is that lookup of the meta data from a NMSetting instance,
  from a NM_TYPE_SETTING or from a NMSettingClass, is now mostly just following
  some pointers. Previously, it always required a hash table lookup.
  Likewise, lookup of a NMMetaSettingInfo from a NMSetting is now just
  following some pointers.

- it reduces duplication. The setting-name is no longer tracked twice,
  but only in the NMMetaSettingInfo. Also, it moves the setting priority
  into NMMetaSettingInfo, it fits better there.